(refactor) Namespace routes for app versioning
